# src/docs/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Lava Rapido API",
    "description": "API do banco de dados",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/veiculos/": {
      "get": {
        "description": "Busca todos os veiculos",
        "responses": {
          "200": {
            "description": "Retorna lista de veiculos",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "veiculo_id": {
                    "type": "number",
                    "example": 2
                  },
                  "cliente_id": {
                    "type": "string",
                    "example": "nome do cliente"
                  },
                  "placa": {
                    "type": "string",
                    "example": "identificacao da placa"
                  },
                  "modelo": {
                    "type": "string",
                    "example": "modelo do veiculo"
                  },
                  "cor": {
                    "type": "string",
                    "example": "cor do veiculo"
                  },
                  "ano": {
                    "type": "number",
                    "example": 2025
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "description": "Cria um veiculo",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "cliente_id": {
                  "type": "number",
                  "example": 1
                },
                "placa": {
                  "type": "string",
                  "example": "identificacao da placa"
                },
                "modelo": {
                  "type": "string",
                  "example": "modelo do veiculo"
                },
                "cor": {
                  "type": "string",
                  "example": "cor do veiculo"
                },
                "ano": {
                  "type": "number",
                  "example": 2025
                }
              },
              "required": [
                "cliente_id",
                "placa",
                "modelo",
                "cor",
                "ano"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Veiculo criado com sucesso",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Veiculo criado com sucesso."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/veiculos/{id}": {
      "get": {
        "description": "Busca um veiculo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna um veiculo",
            "schema": {
              "type": "object",
              "properties": {
                "veiculo_id": {
                  "type": "number",
                  "example": 2
                },
                "cliente_id": {
                  "type": "number",
                  "example": 2
                },
                "placa": {
                  "type": "string",
                  "example": "pma0987"
                },
                "modelo": {
                  "type": "string",
                  "example": "corolla"
                },
                "cor": {
                  "type": "string",
                  "example": "preto"
                },
                "ano": {
                  "type": "number",
                  "example": 2025
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "put": {
        "description": "Edita um veiculo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "cliente_id": {
                  "type": "number",
                  "example": 2
                },
                "placa": {
                  "type": "string",
                  "example": "identificacao da placa"
                },
                "modelo": {
                  "type": "string",
                  "example": "modelo do veiculo"
                },
                "cor": {
                  "type": "string",
                  "example": "cor do veiculo"
                },
                "ano": {
                  "type": "number",
                  "example": 2025
                }
              },
              "required": [
                "cliente_id",
                "placa",
                "modelo",
                "cor",
                "ano"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Veiculo atualizado com sucesso",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Veiculo atualizado com sucesso."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "description": "Deleta um veiculo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "registro deletado",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "success"
                },
                "description": {
                  "type": "string",
                  "example": "Registro deletado com sucesso."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}